swipe deleting

+++ cclasswork020323/app/src/main/java/com/example/cclasswork020323/dao/StudentDao.java
@@ -10,6 +10,6 @@ public interface StudentDao {
     List<Student> findAll();
     Student findById(long id);
     int update(long id, Student student);
-    int deleteById(int id);
+    int deleteById(long id);
 
 }
