+++ cclasswork020323/app/src/main/java/com/example/cclasswork020323/dao/StudentDaoImpl.java
@@ -155,7 +155,7 @@ public class StudentDaoImpl implements StudentDao{
     }
 
     @Override
-    public int deleteById(int id) {
+    public int deleteById(long id) {
 
         SQLiteDatabase sqLiteDatabase = openHelper.getWritableDatabase();
